% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{agrestiCoull}
\alias{agrestiCoull}
\title{Agresti-Coull estimation of uncertainty on CDF}
\usage{
agrestiCoull(X, n)
}
\arguments{
\item{X}{-}

\item{n}{-}
}
\description{
Agresti-Coull estimation of uncertainty on CDF
}
