% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibScores.R
\name{calibScores}
\alias{calibScores}
\title{Calibration, sharpness, consistency and adaptativity scores}
\usage{
calibScores(E, u, X = NULL, nBin = NULL, intrv = NULL)
}
\arguments{
\item{E}{(vector) errors}

\item{u}{(vector) prediction uncertainties}

\item{X}{(vector/matrix) input features / coordinates (default: `NULL`)}

\item{nBin}{(integer) number of bins used for statistics
(default: `sqrt(length(E))`). Superseded by `intrv$nbr` if present.}

\item{intrv}{(list) intervals generated by `genIntervals` (default: `NULL`)}
}
\value{
A list containing z-scores based statistics(Scal, Su, SX and Stot),
  ENCE (enceu, enceX) and UCE (uceu, uceX) calibration errors, the NLL and
  its calibration and sharpness components (nll, calib, sharp).
  The presence and length of SX, enceX and uceX match the presence and
  column number of X.
}
\description{
Calibration, sharpness, consistency and adaptativity scores
}
\examples{
\donttest{
  uE = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E  = rnorm(uE, mean=0, sd=uE)    # Generate errors
  calibScores(E, uE)
}
}
