% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibScores.R
\name{calibScoresProbRef}
\alias{calibScoresProbRef}
\title{Probabilistic reference values for calibration scores}
\usage{
calibScoresProbRef(
  E,
  u,
  X = NULL,
  nBin = NULL,
  intrv = NULL,
  nMC = 1000,
  dist = c("Normal", "Uniform", "Normp4", "Laplace", "T4"),
  quant = FALSE
)
}
\arguments{
\item{E}{(vector) errors}

\item{u}{(vector) prediction uncertainties}

\item{X}{(vector/matrix) input features / coordinates (default: `NULL`)}

\item{nBin}{(integer) number of bins used for statistics
(default: `sqrt(length(E))`). Superseded by `intrv$nbr` if present.}

\item{intrv}{(list) intervals generated by `genIntervals` (default: `NULL`)}

\item{nMC}{(integer) sampling size (default: 1000)}

\item{dist}{(string) model error distribution to generate the reference
values. One of 'Normal' (default), 'Uniform', 'Normp4', 'Laplace' or 'T4'}

\item{quant}{(logical) if TRUE, return 95\% confidence interval limits}
}
\value{
A list containing two lists: meanRefScores is a list of mean scores
  with the same structure as the output of `calibScores` and sdRefScores
  contains the standard deviations with the same structure
}
\description{
Probabilistic reference values for calibration scores
}
\examples{
\donttest{
  uE = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E  = rnorm(uE, mean=0, sd=uE)    # Generate errors
  calibScoresProbRef(E, uE)
}

}
