% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibScores.R
\name{eceFun}
\alias{eceFun}
\title{ENCE and UCE}
\usage{
eceFun(E, uE, intrv)
}
\arguments{
\item{E}{(vector) errors}

\item{intrv}{(list) intervals generated by `genIntervals`}

\item{u}{(vector) prediction uncertainties}
}
\value{
A list with ence and uce values
}
\description{
ENCE and UCE
}
\examples{
\donttest{
  uE = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E  = rnorm(uE, mean=0, sd=uE)    # Generate errors
  intrv = ErrViewLib::genIntervals(uE, 30)
  eceFun(E, uE, intrv)
}
}
