% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBS.R
\name{estBS1}
\alias{estBS1}
\title{Estimate benchmarking statistics and their uncertainty by bootstrapping}
\usage{
estBS1(
  error,
  props = c("mue", "mse", "rmsd", "skew", "kurt", "q95", "q95hd", "P1", "W"),
  nboot = 1000,
  do.sip = TRUE,
  eps = 1,
  seed = 123,
  silent = TRUE
)
}
\arguments{
\item{error}{(list)}

\item{props}{(strings; optional) a vector of statistics to compute.}

\item{nboot}{(numeric; optional) the number of bootstrap repetitions.}

\item{do.sip}{(logical; optional) perform SIP analysis.}

\item{eps}{(numeric; optional) threshold for the estimation of P1.}

\item{seed}{(numeric; optional) seed for RNG.}

\item{silent}{(logical; optional) flag to hide progress messages.}
}
\description{
Estimate benchmarking statistics and their uncertainty by bootstrapping
}
