% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fcor}
\alias{fcor}
\title{Correlation}
\usage{
fcor(X, index = 1:length(X), ...)
}
\arguments{
\item{X}{-}

\item{index}{-}

\item{...}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
