% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fdif}
\alias{fdif}
\title{Difference of statistics}
\usage{
fdif(X, index = 1:nrow(X), fscore, ...)
}
\arguments{
\item{X}{-}

\item{index}{-}

\item{fscore}{-}

\item{...}{-}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
