% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{forder}
\alias{forder}
\title{Order of a series of statistics}
\usage{
forder(data, index = 1:nrow(data), fscore, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
