% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fOrderMSIP}
\alias{fOrderMSIP}
\title{Order of a series of MSIP statistics}
\usage{
fOrderMSIP(data, index = 1:nrow(data), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
