% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTabStat.R
\name{formatUnc}
\alias{formatUnc}
\title{Truncate value and uncertainty to consistent number of digits.}
\usage{
formatUnc(y, uy, numDig = 2)
}
\arguments{
\item{y}{(numeric) value}

\item{uy}{(numeric) uncertainty on `y`}

\item{numDig}{(numeric) number of digits to keep on `uy`}
}
\value{
A list with truncated values of `y` and `uy`.
}
\description{
Truncate value and uncertainty to consistent number of digits.
}
