% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{frank}
\alias{frank}
\title{Ranks of a series of statistics}
\usage{
frank(data, index = 1:nrow(data), fscore, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Auxillary function for bootstrap by 'boot::boot()'
}
