% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fRankMSIP}
\alias{fRankMSIP}
\title{Ranks of a series of MSIP statistics}
\usage{
fRankMSIP(data, index = 1:nrow(data), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Auxillary function for bootstrap by 'boot::boot()'
}
