% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTabStat.R
\name{genpval}
\alias{genpval}
\title{Generalized p-value (Liu & Sing 1997, Wilcox 2012)}
\usage{
genpval(X)
}
\arguments{
\item{X}{(vector) a set of values}
}
\value{
a p-value
}
\description{
Generalized p-value (Liu & Sing 1997, Wilcox 2012)
}
