% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTabStat.R
\name{genTabStat}
\alias{genTabStat}
\title{Generate results table from benchmarking statistics generated
by 'estBS()'}
\usage{
genTabStat(S, comp = TRUE, ref = 0, numDig = 1, units = "a.u.", short = TRUE)
}
\arguments{
\item{S}{-}

\item{comp}{-}

\item{ref}{-}

\item{numDig}{(integer) number of digits to keep for uncertainty display.}

\item{units}{(string) units of the data.}

\item{short}{(logical) use parenthetic notation to display uncertainty.}
}
\value{
A dataframe.
}
\description{
Generate results table from benchmarking statistics generated
by 'estBS()'
}
