% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gimc}
\alias{gimc}
\title{Gini coefficient of mode-centered absolute errors}
\usage{
gimc(X, index = 1:length(X), ...)
}
\arguments{
\item{X}{-}

\item{index}{-}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
