% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gini}
\alias{gini}
\title{Gini coefficient of absolute errors}
\usage{
gini(X, index = 1:length(X), ...)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
