% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdQuantiles.R
\name{hd}
\alias{hd}
\title{Estimate quantiles by Harrell & Davis 1982}
\usage{
hd(x, q = 0.5, na.rm = TRUE)
}
\arguments{
\item{x}{(vector) a set of values}

\item{q}{(numeric) a probability level}

\item{na.rm}{(logical) remove NAs}
}
\value{
A quantile.
}
\description{
Estimate quantiles by Harrell & Davis 1982
}
