% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hrmode}
\alias{hrmode}
\title{Mode of signed errors by half-range estimator}
\usage{
hrmode(X, index = 1:length(X), ...)
}
\arguments{
\item{X}{-}

\item{index}{-}
}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
