% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hsmode}
\alias{hsmode}
\title{Mode of signed errors by half-sample estimator}
\usage{
hsmode(X, index = 1:length(X), ...)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
