% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kurtcs}
\alias{kurtcs}
\title{Robust Kurtosis (Crow & Siddiqui)}
\usage{
kurtcs(X, index = 1:length(X), ...)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
