% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lasym}
\alias{lasym}
\title{Lorenz curve asymmetry of absolute errors}
\usage{
lasym(X, index = 1:length(X), ...)
}
\arguments{
\item{X}{-}

\item{index}{-}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
