% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mad_sd}
\alias{mad_sd}
\title{Median Absolute Deviation (MAD-SD)}
\usage{
mad_sd(X, index = 1:length(X), ...)
}
\arguments{
\item{X}{-}

\item{index}{-}
}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
