% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.R
\name{myGetColors}
\alias{myGetColors}
\title{Generate colors according to smooth rainbow scheme}
\usage{
myGetColors(n, alpha = NULL)
}
\arguments{
\item{n}{(integer) number of colors}

\item{alpha}{(numeric) level of alpha channel in [0,1]}
}
\value{
a color palette
}
\description{
Generate colors according to smooth rainbow scheme
}
