% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibScores.R
\name{nllFun}
\alias{nllFun}
\title{Negative Log Likelihood and components}
\usage{
nllFun(E, u)
}
\arguments{
\item{E}{(vector) errors}

\item{u}{(vector) prediction uncertainties}
}
\value{
A list with the nll and the calibration and sharpness scores
}
\description{
Negative Log Likelihood and components
}
\examples{
\donttest{
  uE = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E  = rnorm(uE, mean=0, sd=uE)    # Generate errors
  nllFun(E, uE)
}
}
