% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{P1}
\alias{P1}
\title{Probability for the absolute errors to be below a threshold}
\usage{
P1(X, index = 1:length(X), eps)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
