% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParallel.R
\name{paraPlot}
\alias{paraPlot}
\title{Title}
\usage{
paraPlot(
  x,
  col = 1,
  lty = 1,
  pch = 16,
  las = las,
  var.label = NULL,
  lab.thresh = 0,
  rescale = FALSE,
  scramble = FALSE,
  outliers = "no",
  outLabCex = 1,
  xlim = c(1, ncol(x)),
  ylim = NULL,
  ylab = "",
  cols_tr2 = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
