% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTabStat.R
\name{pinv}
\alias{pinv}
\title{Probability to have a sign different from the sign of d0}
\usage{
pinv(X, d0)
}
\arguments{
\item{X}{(vector) values to be tested.}

\item{d0}{(numeric) reference value.}
}
\value{
A probability.
}
\description{
Probability to have a sign different from the sign of d0
}
