% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACF.R
\name{plotACF}
\alias{plotACF}
\title{Wrapper for the acf function}
\usage{
plotACF(
  X,
  lag.max = NULL,
  title = "",
  col = 5,
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) a numerical vector}

\item{lag.max}{(integer)}

\item{title}{(string)}

\item{col}{(integer) color index of the CI}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
a plot
}
\description{
Wrapper for the acf function
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  res = plotLZMS(uE, E/uE, ylim = c(0,2))
  plotACF(res$pc)
}
}
