% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAGVZMS.R
\name{plotAGV}
\alias{plotAGV}
\title{Adversarial Group Calibration based on the Scal statistic}
\usage{
plotAGV(
  Z,
  popMin = 50,
  nGroup = 100,
  nMC = 500,
  add = FALSE,
  ylim = NULL,
  col = 6,
  control = TRUE,
  colControl = 2,
  stat = c("max", "mean", "median", "q95"),
  dist = c("Normal", "Uniform", "Normpn", "Normp4", "Laplace", "T4", "Student"),
  df = 4,
  title = "",
  label = 0,
  legend = TRUE,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{Z}{(vector) set of z-score values to be tested}

\item{popMin}{(integer) minimal bin count in an interval}

\item{nGroup}{(integer) number random groups sampled to select worst one}

\item{nMC}{(integer) number of repeats for worst group selection}

\item{add}{(logical) add to previous graph ?}

\item{ylim}{(vector) limits of the y axis}

\item{col}{(integer) color index of main curve}

\item{control}{(logical) estimate AGV for control sample (normal-standard)}

\item{colControl}{(integer) color index of control curve}

\item{dist}{(string) model error distribution to generate the control
values. One of 'Normal' (default), 'Uniform', 'Normpn', Normp4', 'Laplace', 'Tn' or 'T4'}

\item{df}{(integer) degrees of freedom for distributions 'Normpn' and 'Tn'}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{legend}{(logical) add a legend (default: TRUE) ?}

\item{gPars}{(list) graphical parameters}

\item{method}{(string) method used to estimate 95 percent CI on <Z^2>}
}
\value{
Nothing.
}
\description{
Adversarial Group Calibration based on the Scal statistic
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  plotAGV(E/uE)
}
}
