% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAGVZMS.R
\name{plotAGVZMS}
\alias{plotAGVZMS}
\title{Adversarial Group Validation of calibration based on the
fraction of validated confidence intervals of the <Z^2>
statistic}
\usage{
plotAGVZMS(
  Z,
  popMin = 100,
  popMax = length(Z)/10,
  mZ2 = 1,
  nBoot = 1500,
  method = c("bootstrap", "stud", "auto"),
  BSmethod = c("bca", "perc", "basic"),
  nMC = 1000,
  add = FALSE,
  ylim = NULL,
  col = 6,
  verbose = FALSE,
  control = TRUE,
  colControl = 2,
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{Z}{(vector) set of z-score values to be tested}

\item{popMin}{(integer) minimal bin count in an interval}

\item{popMax}{(integer) maximal bin count in an interval}

\item{mZ2}{(numeric) target value for <Z^2> (default `1`)}

\item{nBoot}{(integer) number of bootstrap replicas}

\item{method}{(string) method used to estimate 95 percent CI on <Z^2>}

\item{BSmethod}{(string) bootstrap variant}

\item{nMC}{(integer) number of random intervals per size}

\item{add}{(logical) add to previous graph ?}

\item{ylim}{(vector) limits of the y axis}

\item{col}{(integer) color index of main curve}

\item{verbose}{(logical) print progress messages}

\item{control}{(logical) estimate AGV for control sample (normal-standard)}

\item{colControl}{(integer) color index of control curve}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of worst stats. Mainly used
  for its plotting side effect.
}
\description{
Adversarial Group Validation of calibration based on the
fraction of validated confidence intervals of the <Z^2>
statistic
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  plotAGVZMS(E/uE, method = 'stud')
}
}
