% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBA.R
\name{plotBA}
\alias{plotBA}
\title{Bland-Altman-type plot of error samples}
\usage{
plotBA(data1, data2, ylim = NULL, title = "", gPars = ErrViewLib::setgPars())
}
\arguments{
\item{data1}{-}

\item{data2}{-}

\item{ylim}{-}

\item{title}{-}

\item{gPars}{-}
}
\description{
Bland-Altman-type plot of error samples
}
