% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalCurve.R
\name{plotCalCurve}
\alias{plotCalCurve}
\title{Plot local coverage probabilities to assess calibration and sharpness}
\usage{
plotCalCurve(
  Data,
  corTrend = FALSE,
  fo = NA,
  CImeth = c("eq", "pred", "dist"),
  prob = seq(0.005, 0.995, by = 0.005),
  dist = c("norm", "t"),
  shape = 2,
  valid = c("kfold", "loo"),
  nFold = 10,
  nRepeat = 10,
  score = TRUE,
  binomCI = c("wilson", "wilsoncc", "clopper-pearson", "agresti-coull", "jeffreys"),
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{Data}{(data.frame) dataframe with predictor(s) and reference}

\item{corTrend}{(logical) flag to correct trend}

\item{fo}{(formula) trend correction formula}

\item{CImeth}{(string) method to estimate CI limits}

\item{prob}{(vector) a set of coverage probabilities to test}

\item{dist}{(string) a distribution}

\item{shape}{(numeric) shape parameter (> 0, maybe non-integer).}

\item{valid}{(string) a cross-validation method: "kfold" (default) or "loo"}

\item{nFold}{(integer) number of folds (default: 10)}

\item{nRepeat}{(integer) number of repeats for k-fold cross-validation}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}

\item{uP}{(vector) a set of prediction uncertainties}

\item{mycols}{(vector) a set of color indexes to gPars colors}

\item{plot}{(logical) plot the results}
}
\value{
Invisibly returns a list of calibration statistics. Mainly used
  for its plotting side effect.
}
\description{
Plot local coverage probabilities to assess calibration and sharpness
}
