% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalCurve.R
\name{plotCalibrationQ}
\alias{plotCalibrationQ}
\title{Plot calibration curve from quantiles table}
\usage{
plotCalibrationQ(
  E,
  Q,
  cond = NULL,
  nBin = NULL,
  intrv = NULL,
  equiPop = TRUE,
  popMin = 30,
  logBin = TRUE,
  score = TRUE,
  plot = TRUE,
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{E}{(vector) a vector of errors}

\item{cond}{(vector) a vector of values on which conditional calibration curves are designed}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{intrv}{(object) intervals generated by `genIntervals` (default: `NULL`)}

\item{equiPop}{(logical) generate intervals  with equal bin counts
(default: `equiPop = TRUE`)}

\item{popMin}{(integer) minimal bin count in an interval}

\item{logBin}{(logical) if `equiPop = FALSE`, one can choose between
equal range intervals, or equal log-range intervals
(default `logBin = TRUE`)}

\item{score}{(logical) evaluate the calibration scores (default: `TRUE`)}

\item{plot}{(logical) plot the results (default: `TRUE`)}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of calibration statistics. Mainly used
  for its plotting side effect.
}
\description{
Plot calibration curve from quantiles table
}
