% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfidence.R
\name{plotCC}
\alias{plotCC}
\title{Plot confidence curve for (uE,E) set. Interface to plotConfidence
with different default arguments.}
\usage{
plotCC(
  E,
  uE,
  statS = c("RMSE", "RMSD", "MAE", "Q95"),
  normalize = FALSE,
  oracle = FALSE,
  probref = TRUE,
  conf_probref = TRUE,
  dist_probref = c("Normal", "Uniform", "Normp4", "Laplace", "T4"),
  rep_probref = 500,
  plot = TRUE,
  score = TRUE,
  confStat = score,
  col = 6,
  add = FALSE,
  xlab = "k\% discarded",
  xlim = NULL,
  ylim = NULL,
  title = NULL,
  label = 0,
  showUk = FALSE,
  showLegend = TRUE,
  unit = "",
  legend = NULL,
  legLoc = "bottomleft",
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{E}{(vector) prediction uncertainty, uE, or predicted value, V}

\item{uE}{(vector) error or z-score}

\item{statS}{(string) statistic to use. One of 'RMSE' (default), 'MAE' or 'Q95'}

\item{normalize}{(logical) use normalized statistic (default: FALSE)}

\item{oracle}{(logical) plot oracle curve (default: FALSE)}

\item{probref}{(logical) plot probabilistic reference (probref) curve (default: TRUE)}

\item{conf_probref}{(logical) plot confidence band around probref curve (default: TRUE)}

\item{dist_probref}{(string) model error distribution to generate the reference curve. One of 'Normal' (default), 'Uniform', 'Normp4', 'Laplace' or 'T4'}

\item{rep_probref}{(integer) sampling repetitions for normal curve (default = 500)}

\item{plot}{(function) produce plot ?}

\item{score}{(logical) estimate DFPR and/or AUCO}

\item{confStat}{(logical) estimate DFPR and/or AUCO}

\item{col}{(integer) color index for the curve (default: 6)}

\item{add}{(logical) add confidence curve to previous plot (default: FALSE)}

\item{xlab}{(string) x axis label (default: 'k \% discarded)}

\item{xlim}{(vector) limits of the x axis (default: NULL)}

\item{ylim}{(vector) limits of the y axis (default: NULL)}

\item{title}{(string) a title to display above the plot (default: '')}

\item{label}{(integer) index of letter for subplot tag (default: 0)}

\item{showUk}{(logical) plot secondary axis with u_k values; supersedes `title`}

\item{showLegend}{(logical) display legend (default: TRUE)}

\item{unit}{(string) unit string to add to ylab}

\item{legend}{(string) legend for the dataset (default: NULL)}

\item{legLoc}{(string) location of legend (see \link[grDevices]{xy.coord}) (default: 'bottomleft')}

\item{gPars}{(list) graphical parameters (default: ErrViewLib::setgPars())}
}
\value{
Plot confidence curve for (E,uE) set
}
\description{
Plot confidence curve for (uE,E) set. Interface to plotConfidence
with different default arguments.
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)  # Generate errors
  plotCC(E, uE, statS = 'MAE')
}
}
