% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfidence.R
\name{plotCCVal}
\alias{plotCCVal}
\title{Plot confidence curve for (uE,E) set. New version with improved validation.}
\usage{
plotCCVal(
  E,
  uE,
  stat = ErrViewLib::rmse,
  probref = FALSE,
  rep_probref = 1000,
  showUk = FALSE,
  plot = TRUE,
  score = FALSE,
  nBoot = 5000,
  parallel = FALSE,
  cl = NULL,
  col = 5,
  type = c("l", "p"),
  add = FALSE,
  xlab = "k\% discarded",
  xlim = NULL,
  ylab = "RMSE",
  ylim = NULL,
  title = NULL,
  label = 0,
  showLegend = TRUE,
  legend = NULL,
  legLoc = "bottomleft",
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{E}{(vector) prediction uncertainty, uE, or predicted value, V}

\item{uE}{(vector) error or z-score}

\item{probref}{(logical) plot probabilistic reference (probref) curve (default: FALSE)}

\item{rep_probref}{(integer) sampling repetitions for normal curve (default = 1000)}

\item{showUk}{(logical) plot secondary axis with u_k values; supersedes `title`}

\item{plot}{(function) produce plot ?}

\item{score}{(logical) estimate confidence score ? This requires a bootstrapped CI on the CC and might take some time.}

\item{nBoot}{(integer) nb of bootstrap repeats to estimate CI on confidence curve (default = 5000)}

\item{parallel}{(logical) use parallelized bootstrap ? (default = FALSE)}

\item{cl}{(object) cluster for parallel computing, which is used when parallel = TRUE. If parallel = TRUE and cl = NULL, then the cluster is defined as cl = makeCluster(detectCores()). (default = NULL)}

\item{col}{(integer) color index for the curve (default: 6)}

\item{add}{(logical) add confidence curve to previous plot (default: FALSE)}

\item{xlab}{(string) x axis label (default: 'k \% discarded)}

\item{xlim}{(vector) limits of the x axis (default: NULL)}

\item{ylim}{(vector) limits of the y axis (default: NULL)}

\item{title}{(string) a title to display above the plot (default: '')}

\item{label}{(integer) index of letter for subplot tag (default: 0)}

\item{showLegend}{(logical) display legend (default: TRUE)}

\item{legend}{(string) legend for the dataset (default: NULL)}

\item{legLoc}{(string) location of legend (see \link[grDevices]{xy.coord}) (default: 'bottomleft')}

\item{gPars}{(list) graphical parameters (default: ErrViewLib::setgPars())}

\item{unit}{(string) unit string to add to ylab}
}
\value{
Plot confidence curve for (E,uE) set
}
\description{
Plot confidence curve for (uE,E) set. New version with improved validation.
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)  # Generate errors
  plotCCVal(E, uE)
}
}
