% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot dendrograms for hierarchical clusering of systems or methods}
\usage{
plotClusters(
  X,
  type = c("systems", "methods"),
  method = c("complete", "single"),
  xlim = NULL,
  ylim = NULL,
  cex.lab = 1,
  gPars
)
}
\arguments{
\item{X}{a numeric matrix of error values}

\item{type}{(optional) a string describing if "systems" or "methods"
are to be clustered. Default : "systems".}

\item{method}{(optional) a string describing the hierarchical
clustering method. Default : "complete".}

\item{xlim}{(optional) a 2-vector of limits for the x-axis.}

\item{ylim}{(optional) a 2-vector of limits for the y-axis.}

\item{cex.lab}{(optional) a scale factor for the size of labels.}

\item{gPars}{a list of graphical parameters. See \linl{setgPars}}
}
\value{
a plot
}
\description{
Plot dendrograms for hierarchical clusering of systems or methods
}
