% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfidence.R
\name{plotConfidence}
\alias{plotConfidence}
\title{Plot confidence curve for (uE,E) set}
\usage{
plotConfidence(
  E,
  uE,
  stat = ErrViewLib::mue,
  oracle = TRUE,
  ideal = FALSE,
  conf_ideal = FALSE,
  dist_ideal = "Normal",
  rep_ideal = 100,
  col = 2,
  type = c("l", "p"),
  add = FALSE,
  xlab = "k\% discarded",
  xlim = NULL,
  ylab = "MAE / MAE0",
  ylim = NULL,
  title = NULL,
  label = 0,
  legend = NULL,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{E}{(vector) prediction uncertainty, uE, or predicted value, V}

\item{uE}{(vector) error or z-score}

\item{stat}{(function) statistic to use}

\item{oracle}{(logical) plot Oracle curve}

\item{ideal}{(logical) plot Ideal curve}

\item{conf_ideal}{(logical) plot confidence band around Ideal curve}

\item{dist_ideal}{(string) model error distribution to generate Ideal curve}

\item{rep_ideal}{(integer) sampling repetitions for normal curve}

\item{col}{(integer) color index for the curve}

\item{type}{(string) curve type: line ('l') or points ('p')}

\item{add}{(logical) add confidence curve to previous plot}

\item{xlab}{(string) x axis label}

\item{xlim}{(vector) limits of the x axis}

\item{ylab}{(string) y axis label}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{legend}{(string) legend for the dataset}

\item{gPars}{(list) graphical parameters}
}
\value{
Plot confidence curve for (E,uE) set
}
\description{
Plot confidence curve for (uE,E) set
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)  # Generate errors
  plotConfidence(E,uE)
}

}
