% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistHist.R
\name{plotDistHist}
\alias{plotDistHist}
\title{Scatter plot of an error set with marginal histogram}
\usage{
plotDistHist(
  x,
  y,
  uy = NULL,
  nclass = NULL,
  xlab = "x",
  ylab = "y",
  logX = FALSE,
  plotGauss = FALSE,
  outLiers = FALSE,
  p = 0.9,
  labels = 1:length(x),
  select = NULL,
  main = NULL,
  plotReg = TRUE,
  plotConf = FALSE,
  degree = 0,
  colPoints = NULL,
  plotBA = FALSE,
  plotBAci = FALSE,
  topMar = 1.6,
  xlim = range(x),
  ylim = range(y),
  scaleLegBA = 0.75,
  scalePoints = 0.75,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{x}{-}

\item{y}{-}

\item{uy}{-}

\item{nclass}{-}

\item{xlab}{-}

\item{ylab}{-}

\item{logX}{-}

\item{plotGauss}{-}

\item{outLiers}{-}

\item{p}{-}

\item{labels}{-}

\item{select}{-}

\item{main}{-}

\item{plotReg}{-}

\item{plotConf}{-}

\item{degree}{-}

\item{colPoints}{-}

\item{plotBA}{-}

\item{plotBAci}{-}

\item{topMar}{-}

\item{xlim}{-}

\item{ylim}{-}

\item{scaleLegBA}{-}

\item{gPars}{-}
}
\value{

}
\description{
Scatter plot of an error set with marginal histogram
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  V   = 1:length(E)
  plotDistHist(V, E, uE, xlab = 'V', ylab='E',ylim = c(-12,12))
}
}
