% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvsPU.R
\name{plotEvsPU}
\alias{plotEvsPU}
\title{Plot error distribution or extremal errors vs prediction uncertainty}
\usage{
plotEvsPU(
  x,
  y,
  type = c("points", "ext"),
  xlab = "Pred. unc.",
  ylab = "Error",
  logX = FALSE,
  title = NULL,
  colPoints = NULL,
  xlim = NULL,
  ylim = NULL,
  scalePoints = 0.75,
  nBin = NULL,
  slide = NULL,
  gPars
)
}
\arguments{
\item{x}{(vector) prediction uncertainty}

\item{y}{(vector) error}

\item{type}{(string) plot type. Default 'points'}

\item{xlab}{(string) x axis label. Default 'Pred. unc.'}

\item{ylab}{(string) y axis label. Default 'Error'}

\item{logX}{(logical) log-transform x}

\item{title}{(string) a title to display above the plot}

\item{xlim}{(vector) limits of the x axis}

\item{ylim}{(vector) limits of the y axis}

\item{scalePoints}{(numeric) scale factor for points size}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{slide}{(logical) use sliding window for subsetting (X,Z)}

\item{gPars}{(list) graphical parameters}

\item{label}{(integer) index of letter for subplot tag}
}
\value{
Plots y vs x or extremal values of abs(y) vs x.
}
\description{
Plot error distribution or extremal errors vs prediction uncertainty
}
