% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGLAC.R
\name{plotGiniVsLAC}
\alias{plotGiniVsLAC}
\title{Plot Gini vs. LAC coefs}
\usage{
plotGiniVsLAC(
  X,
  title = "",
  show.norm = FALSE,
  show.leg = TRUE,
  col.index = 1:ncol(X),
  label = 0,
  leg.lwd = 2,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{-}

\item{title}{-}

\item{show.norm}{-}

\item{show.leg}{-}

\item{col.index}{-}

\item{label}{-}

\item{leg.lwd}{-}

\item{gPars}{-}
}
\description{
Plot Gini vs. LAC coefs
}
