% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLCP.R
\name{plotLCP}
\alias{plotLCP}
\title{Plot local coverage probabilities to assess calibration and sharpness}
\usage{
plotLCP(
  E,
  U,
  ordX = NULL,
  logX = FALSE,
  prob = c(0.95),
  nBin = NULL,
  binomCI = c("wilson", "wilsoncc", "clopper-pearson", "agresti-coull", "jeffreys"),
  plot = TRUE,
  slide = NULL,
  mycols = 1:length(prob),
  xlab = "Calculated value",
  xlim = NULL,
  ylim = c(0, 1),
  title = "",
  legLoc = "bottom",
  legNcol = 3,
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{E}{(vector) Errors}

\item{U}{(matrix) Prediction uncertainties}

\item{ordX}{(vector) set of abscissas to order sample}

\item{logX}{(logical) log-transform abscissas}

\item{prob}{(vector) a set of coverage probabilities for the PUs}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{binomCI}{(string) name of method to estimate Binomial Proportion CI}

\item{plot}{(logical) plot the results}

\item{slide}{(logical) use sliding window}

\item{mycols}{(vector) a set of color indexes to gPars colors}

\item{xlab}{(string) abscissa label}

\item{xlim}{(vector) range for abscissa}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{legLoc}{(string) location of legend (see \link[grDevices]{xy.coord})}

\item{legNcol}{(integer) number of columns for legend}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LCP results. Mainly used
  for its plotting side effect.
}
\description{
Plot local coverage probabilities to assess calibration and sharpness
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  plotLCP(E, cbind(0.32*uE, uE, 1.96*uE), prob=c(0.25,0.67,0.95), ordX = uE)
}
}
