% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLorenz.R
\name{plotLorenz}
\alias{plotLorenz}
\title{Plot of a set of Lorenz curves}
\usage{
plotLorenz(
  X,
  var = FALSE,
  title = " Lorenz curves",
  show.norm = FALSE,
  show.leg = TRUE,
  col.index = 1:ncol(X),
  identity.grid = FALSE,
  label = 0,
  leg.lwd = 2,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{-}

\item{var}{-}

\item{title}{-}

\item{show.norm}{-}

\item{show.leg}{-}

\item{col.index}{-}

\item{identity.grid}{-}

\item{label}{-}

\item{leg.lwd}{-}

\item{gPars}{-}
}
\value{

}
\description{
Plot of a set of Lorenz curves
}
