% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistHist.R
\name{plotlyDistHist}
\alias{plotlyDistHist}
\title{Scatter plot of an error set with marginal histogram (plotly version)}
\usage{
plotlyDistHist(
  x,
  y,
  uy = NULL,
  nclass = NULL,
  xlab = "x",
  ylab = "y",
  plotGauss = FALSE,
  outLiers = FALSE,
  p = 0.9,
  labels = 1:length(x),
  select = NULL,
  main = NULL,
  plotReg = TRUE,
  plotConf = FALSE,
  degree = 0,
  plotBA = FALSE,
  plotBAci = FALSE,
  xlim = range(x),
  ylim = range(y),
  scaleLegBA = 0.75,
  scalePoints = 0.75,
  gPars
)
}
\arguments{
\item{x}{-}

\item{y}{-}

\item{uy}{-}

\item{nclass}{-}

\item{xlab}{-}

\item{ylab}{-}

\item{plotGauss}{-}

\item{outLiers}{-}

\item{p}{-}

\item{labels}{-}

\item{select}{-}

\item{main}{-}

\item{plotReg}{-}

\item{plotConf}{-}

\item{degree}{-}

\item{plotBA}{-}

\item{plotBAci}{-}

\item{xlim}{-}

\item{ylim}{-}

\item{scaleLegBA}{-}

\item{gPars}{-}
}
\description{
Scatter plot of an error set with marginal histogram (plotly version)
}
