% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZISD.R
\name{plotLZISD}
\alias{plotLZISD}
\title{Plot local inverse of z-score standard deviation 1/SD(Z)}
\usage{
plotLZISD(
  X,
  Z,
  sdZ = 1,
  logX = FALSE,
  nBin = NULL,
  intrv = NULL,
  equiPop = TRUE,
  popMin = 30,
  logBin = TRUE,
  plot = TRUE,
  slide = FALSE,
  method = c("cho", "bootstrap"),
  BSmethod = c("bca", "perc", "basic"),
  nBoot = 500,
  xlab = "Calculated value",
  xlim = NULL,
  ylim = NULL,
  title = "",
  score = FALSE,
  add = FALSE,
  col = 5,
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) abscissae of the Z values}

\item{Z}{(vector) set of z-score values to be tested}

\item{sdZ}{(numeric) target value for 1/SD(Z) (default `1`)}

\item{logX}{(logical) log-transform X}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{intrv}{(object) intervals generated by `genIntervals` (default: `NULL`)}

\item{equiPop}{(logical) generate intervals  with equal bin counts
(default: `TRUE`)}

\item{popMin}{(integer) minimal bin count in an interval}

\item{logBin}{(logical) if `equiPop = FALSE`, one can choose between
equal range intervals, or equal log-range intervals
(default `TRUE`)}

\item{plot}{(logical) plot the results}

\item{slide}{(logical) use sliding window for subsetting (X,Z)}

\item{method}{(string) method used to estimate 95 percent CI on 1/SD(Z)}

\item{BSmethod}{(string) bootstrap variant}

\item{nBoot}{(integer) number of bootstrap replicas}

\item{xlab}{(string) X axis label}

\item{xlim}{(vector) min and max values of X axis}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{score}{(logical) estimate calibration stats (default: `FALSE`)}

\item{add}{(logical) add to previous graph ?}

\item{col}{(integer) color index of curve to add}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LZISD results. Mainly used
  for its plotting side effect.
}
\description{
Plot local inverse of z-score standard deviation 1/SD(Z)
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  plotLZISD(uE, E/uE, method = 'cho', ylim = c(0,2))
}
}
