% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZV.R
\name{plotLZV}
\alias{plotLZV}
\title{Plot local z-score variance to assess calibration and tightness}
\usage{
plotLZV(
  X,
  Z,
  varZ = 1,
  logX = FALSE,
  nBin = NULL,
  plot = TRUE,
  slide = NULL,
  method = c("bootstrap", "cho", "chisq"),
  BSmethod = c("bca", "perc", "basic"),
  nBoot = 1500,
  xlab = "Calculated value",
  xlim = NULL,
  ylim = NULL,
  title = "",
  add = FALSE,
  col = 5,
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) abscissae of the Z values}

\item{Z}{(vector) set of z-score values to be tested}

\item{logX}{(logical) log-transform X}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{plot}{(logical) plot the results}

\item{slide}{(logical) use sliding window for subsetting (X,Z)}

\item{method}{(string) method used to estimate 95 percent CI on Var(Z)}

\item{BSmethod}{(string) bootstrap variant}

\item{nBoot}{(integer) number of bootstrap replicas}

\item{xlab}{(string) X axis label}

\item{xlim}{(vector) min and max values of X axis}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{add}{(logical) add to previous graph ?}

\item{col}{(interger) color index of curve to add}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LZV results. Mainly used
  for its plotting side effect.
}
\description{
Plot local z-score variance to assess calibration and tightness
}
