% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParallel.R
\name{plotParallel}
\alias{plotParallel}
\title{Title}
\usage{
plotParallel(
  X,
  maxPoints = nrow(X),
  labels = NULL,
  lab.thresh = 0,
  colors = NULL,
  rescale = TRUE,
  scramble = FALSE,
  outliers = FALSE,
  units = "a.u.",
  ylab = "Errors",
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Title
}
