% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot Principal Component Analysis results}
\usage{
plotPCA(
  X,
  labels,
  xlim = NULL,
  ylim = NULL,
  cex.lab = 1,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{a numeric matrix}

\item{labels}{a strings vector with the names of the data points}

\item{xlim}{(optional) a 2-vector of limits for the x-axis}

\item{ylim}{(optional) a 2-vector of limits for the y-axis}

\item{cex.lab}{(optional) character expansion factor used for
labelling the points.}

\item{gPars}{a list of graphical parameters. See \link{setgPars}.}
}
\value{
a plot
}
\description{
Plot Principal Component Analysis results
}
