% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPdf.R
\name{plotPdf}
\alias{plotPdf}
\title{Plot of a set of PDFs}
\usage{
plotPdf(
  X,
  absErrors = TRUE,
  adjust = 1,
  xlab = NULL,
  xmin = 0,
  xmax = NULL,
  title = "",
  show.leg = TRUE,
  show.MSE = FALSE,
  show.Q95 = TRUE,
  Q.algo = "HD",
  col.index = 1:ncol(X),
  weights = NULL,
  units = "a.u.",
  label = 0,
  leg.lwd = 30,
  fill = FALSE,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Plot of a set of PDFs
}
