% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPIT.R
\name{plotPIT}
\alias{plotPIT}
\title{Plot PIT histograms}
\usage{
plotPIT(R, C, uC, col = 5, breaks = 10, title = "", label = 0, gPars)
}
\arguments{
\item{R}{(vector) a set of reference values}

\item{C}{(vector) a set of predicted values}

\item{uC}{(vector) a set of prediction uncertainties}

\item{col}{(integer) a color number for the histigram}

\item{breaks}{(integer) number of breaks in the histogram}

\item{title}{(string) a title for the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
a plot
}
\description{
Plot PIT histograms
}
