% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRankMat.R
\name{plotRankMat}
\alias{plotRankMat}
\title{Title}
\usage{
plotRankMat(
  E,
  score = "mue",
  type = "levels",
  method = "square",
  nMC = 1000,
  cex.lab = 1,
  show.main = TRUE,
  offset = 0.7,
  M = nrow(E),
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Title
}
