% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSIPMat.R
\name{plotSIPMat}
\alias{plotSIPMat}
\title{Plot of Systematic Improvement Probability matrix}
\usage{
plotSIPMat(
  X,
  method = "circle",
  order = TRUE,
  main = "",
  label = 0,
  cex.lab = 1,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Interface to \code{corrplot::corrplot()}
}
