% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStratZV.R
\name{plotStratRCE}
\alias{plotStratRCE}
\title{Plot local RCE to assess calibration and tightness
for stratified conditioning variables}
\usage{
plotStratRCE(
  X,
  uE,
  E,
  aggregate = TRUE,
  popMin = 30,
  plot = TRUE,
  nBoot = 500,
  xlab = "Conditioning variable",
  ylim = NULL,
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) abscissae of the Z values}

\item{aggregate}{(logical) aggregate contiguous strata smaller than popMin}

\item{popMin}{(integer) minimal count in a stratum}

\item{plot}{(logical) plot the results}

\item{nBoot}{(integer) number of bootstrap replicas}

\item{xlab}{(string) X axis label}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}

\item{Z}{(vector) set of z-score values to be tested}

\item{varZ}{(numeric) target value for Var(Z) (default `1`)}

\item{method}{(string) method used to estimate 95 percent CI on Var(Z)}

\item{BSmethod}{(string) bootstrap variant}
}
\value{
Invisibly returns a list of RCE results. Mainly used
  for its plotting side effect.
}
\description{
Plot local RCE to assess calibration and tightness
for stratified conditioning variables
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  X   = signif(uE,1)
  plotStratRCE(X, uE, E, ylim = c(-1,1))
}
}
