% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStratZV.R
\name{plotStratZMS}
\alias{plotStratZMS}
\title{Plot local <Z^2>
for stratified conditioning variables}
\usage{
plotStratZMS(
  X,
  Z,
  mZ2 = 1,
  aggregate = TRUE,
  popMin = 30,
  plot = TRUE,
  xlab = "Conditioning variable",
  ylim = NULL,
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) conditioning variable}

\item{Z}{(vector) set of z-score values to be tested}

\item{mZ2}{(numeric) target value for <Z^2> (default: 1)}

\item{aggregate}{(logical) aggregate contiguous strata smaller than popMin}

\item{popMin}{(integer) minimal count in a stratum}

\item{plot}{(logical) plot the results}

\item{xlab}{(string) X axis label}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LZV results. Mainly used
  for its plotting side effect.
}
\description{
Plot local <Z^2>
for stratified conditioning variables
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  X   = signif(uE,1)
  plotStratZMS(X, E/uE, ylim = c(0, 2))
}
}
