% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStratZV.R
\name{plotStratZV}
\alias{plotStratZV}
\title{Plot local z-score variance to assess calibration and tightness
for stratified conditioning variables}
\usage{
plotStratZV(
  X,
  Z,
  varZ = 1,
  aggregate = TRUE,
  popMin = 30,
  greedy = TRUE,
  plot = TRUE,
  method = c("cho", "bootstrap", "chisq", "auto"),
  BSmethod = c("bca", "perc", "basic"),
  nBoot = 5000,
  parallel = FALSE,
  xlab = "Conditioning variable",
  ylim = NULL,
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) abscissae of the Z values}

\item{Z}{(vector) set of z-score values to be tested}

\item{varZ}{(numeric) target value for Var(Z) (default `1`)}

\item{aggregate}{(logical) aggregate contiguous strata smaller than popMin (default: TRUE)}

\item{popMin}{(integer) minimal count in a stratum}

\item{greedy}{(logical) use greedy algorithm to merge strata (default: TRUE)}

\item{plot}{(logical) plot the results}

\item{method}{(string) method used to estimate 95 percent CI on Var(Z)}

\item{BSmethod}{(string) bootstrap variant}

\item{nBoot}{(integer) number of bootstrap replicas}

\item{parallel}{(logical) parallelized bootstrap (default: `FALSE`)}

\item{xlab}{(string) X axis label}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LZV results. Mainly used
  for its plotting side effect.
}
\description{
Plot local z-score variance to assess calibration and tightness
for stratified conditioning variables
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  X   = signif(uE,1)
  plotStratZV(X, E/uE, method = 'cho', ylim = c(0,2))
}
}
