% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcoverage.R
\name{predIp}
\alias{predIp}
\title{Auxillary function for plotPcoverage}
\usage{
predIp(
  Learn,
  Test,
  corTrend = FALSE,
  fo = NA,
  prob = c(0.5, 0.75, 0.95),
  CImeth = c("eq", "pred", "dist"),
  dist = c("norm", "t"),
  shape = 2
)
}
\arguments{
\item{corTrend}{-}

\item{fo}{-}

\item{prob}{-}

\item{CImeth}{-}

\item{cLearn}{-}

\item{rLearn}{-}

\item{cTest}{-}

\item{rTest}{-}
}
\value{

}
\description{
Auxillary function for plotPcoverage
}
