% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{q95}
\alias{q95}
\title{95th Quantile of absolute errors (Q95)}
\usage{
q95(X, index = 1:length(X), ...)
}
\arguments{
\item{X}{-}

\item{index}{-}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
