% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{q95hd}
\alias{q95hd}
\title{95th Quantile of absolute errors using the HD algorithm (Q95HD)}
\usage{
q95hd(X, index = 1:length(X))
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxillary function for bootstrap by 'boot::boot()'
}
