% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankBS.R
\name{rankBS}
\alias{rankBS}
\title{N-out of-N Bootstrap of order statistics}
\usage{
rankBS(E, score = "mue", nMC = 1000)
}
\arguments{
\item{E}{-}

\item{score}{-}

\item{nMC}{-}
}
\value{

}
\description{
N-out of-N Bootstrap of order statistics
}
