% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankBS.R
\name{rankBS2}
\alias{rankBS2}
\title{M-out of-N Bootstrap of order statistics}
\usage{
rankBS2(E, score = "mue", nMC = 1000, M = nrow(E))
}
\arguments{
\item{E}{-}

\item{score}{-}

\item{nMC}{-}

\item{M}{-}
}
\value{

}
\description{
M-out of-N Bootstrap of order statistics
}
