% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rce}
\alias{rce}
\title{Relative Calibration Error (RCE)}
\usage{
rce(X, indx = 1:nrow(X))
}
\arguments{
\item{X}{(matrix) a 2-columns matrix containing the uncertainties and the errors}

\item{index}{-}
}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
