% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmsd}
\alias{rmsd}
\title{Root-Mean Squared Deviation (RMSD)}
\usage{
rmsd(X, index = 1:length(X))
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxillary function for bootstrap by 'boot::boot()'
}
