% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfidence.R
\name{Sconf}
\alias{Sconf}
\title{Auxilliary function for plotConfidence}
\usage{
Sconf(X, pcVec = 0:99, stat = ErrViewLib::mue, normalize = TRUE)
}
\arguments{
\item{X}{(vector) errors}

\item{pcVec}{(vector) percentages}

\item{stat}{(function) statistic}

\item{normalize}{(logical) normalize statistic}
}
\value{
Vector of CC stats
}
\description{
Auxilliary function for plotConfidence
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)  # Generate errors
  plotConfidence(E,uE)
}
}
