% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelDiag.R
\name{sdZCI}
\alias{sdZCI}
\title{Estimate statistics of the SD of a sample}
\usage{
sdZCI(
  Z,
  method = c("bootstrap", "cho"),
  CImethod = c("bca", "perc", "basic"),
  nBoot = 1500,
  level = 0.95,
  parallel = c("no", "multicore")
)
}
\arguments{
\item{Z}{(vector) a data sample}

\item{method}{(string) one of 'bootstrap' (default) and 'cho'}

\item{CImethod}{(string) one of 'bca' (default), 'perc' and 'basic'
for the CI estimation algorithm from bootstrap sample}

\item{nBoot}{(integer) number of bootstrap repeats}

\item{level}{(numeric) a probability level for CI (default: 0.95)}

\item{parallel}{(string) one of 'no' (default) and 'multicore'}
}
\value{
A list containing the mean, sd and ci for the SD
    of the Z sample
}
\description{
Estimate statistics of the SD of a sample
}
