% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgPars.R
\name{setgPars}
\alias{setgPars}
\title{Set default graphical parameters}
\usage{
setgPars(type = c("plot", "publish", "shiny"), gPars = list())
}
\arguments{
\item{type}{(string) plot type. Default: 'plot'. Other options are 'shiny' and 'publish'.}

\item{gPars}{(list) list of graphical parameters to supersede defaults}
}
\value{

}
\description{
Set default graphical parameters
}
