% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{skewgm}
\alias{skewgm}
\title{Robust Skewness (SkewGM)}
\usage{
skewgm(X, index = 1:length(X), ...)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
