% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStratZV.R
\name{stratAgg}
\alias{stratAgg}
\title{Aggregate a stratified sample to avoid strata with counts smaller than a threshold}
\usage{
stratAgg(X, popMin = 50, greedy = TRUE)
}
\arguments{
\item{X}{(vector) stratified sample to be aggregated}

\item{popMin}{(integer) minimal count in a stratum}

\item{greedy}{(logical) use greedy algorithm to merge strata (default: TRUE)}
}
\value{
a list containing the new X vector and the values and counts of the new strata
}
\description{
Aggregate a stratified sample to avoid strata with counts smaller than a threshold
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Uncertainty
  X   = signif(uE,1)                # Stratify uE
  ST  = stratAgg(X)
  print(str(ST))
}
}
