% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankBS.R
\name{tabRank}
\alias{tabRank}
\title{Results table for ranking probabilities}
\usage{
tabRank(E, scores = c("mue", "q95hd"), nMC = 1000)
}
\arguments{
\item{nMC}{}
}
\value{

}
\description{
Results table for ranking probabilities
}
