% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibScores.R
\name{tabScoresRef}
\alias{tabScoresRef}
\title{Table of calibration scores with probabilistic references}
\usage{
tabScoresRef(
  E,
  u,
  X = NULL,
  nBin = NULL,
  intrv = NULL,
  nMC = 1000,
  dist = c("Normal", "Uniform", "Normp4", "Laplace", "T4"),
  rawUnc = FALSE
)
}
\arguments{
\item{E}{(vector) errors}

\item{u}{(vector) prediction uncertainties}

\item{X}{(vector/matrix) input features / coordinates (default: `NULL`)}

\item{nBin}{(integer) number of bins used for statistics
(default: `sqrt(length(E))`). Superseded by `intrv$nbr` if present.}

\item{intrv}{(list) intervals generated by `genIntervals` (default: `NULL`)}

\item{nMC}{(integer) sampling size (default: 1000)}

\item{dist}{(string) model error distribution to generate the reference
values. One of 'Normal' (default), 'Uniform', 'Normp4', 'Laplace' or 'T4'}
}
\value{
A table with two lines or three if rawUnc is TRUE.
}
\description{
Table of calibration scores with probabilistic references
}
\examples{
\donttest{
  uE = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E  = rnorm(uE, mean=0, sd=uE)    # Generate errors
  tabScoresRef(E, uE)
}
}
