% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCovIntProb.R
\name{testCovIntProb}
\alias{testCovIntProb}
\title{Tests the fraction of intervals in agreement with target statistics
following LCP or LZV/LZISD analysis}
\usage{
testCovIntProb(stats, prob = 0.95, silent = FALSE)
}
\arguments{
\item{stats}{An object returned by a LCP or LZV/LZISD analysis}

\item{prob}{(numeric) the probability levels of the intervals to be tested}

\item{silent}{(logical) should the function run silently ?}
}
\value{
Invisibly returns a list with binomial test results
}
\description{
Tests the fraction of intervals in agreement with target statistics
following LCP or LZV/LZISD analysis
}
