% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZV.R
\name{varvar}
\alias{varvar}
\title{Var(Var(Z)) by moments formula from Cho2008}
\usage{
varvar(Z)
}
\arguments{
\item{Z}{(vector) a data sample}
}
\value{
The variance on the sample variance
}
\description{
Var(Var(Z)) by moments formula from Cho2008
}
