% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZV.R
\name{varZCI}
\alias{varZCI}
\title{Estimate statistics of the variance of a sample}
\usage{
varZCI(
  Z,
  method = c("bootstrap", "cho", "chisq"),
  CImethod = c("bca", "perc", "basic"),
  nBoot = 1500,
  level = 0.95,
  parallel = c("no", "multicore")
)
}
\arguments{
\item{parallel}{}
}
\value{

}
\description{
Estimate statistics of the variance of a sample
}
