% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdQuantiles.R
\name{vhd}
\alias{vhd}
\title{Quantiles by Harrell & Davis 1982 for a vector of probabilities}
\usage{
vhd(x, p = c(0.025, 0.975), na.rm = TRUE)
}
\arguments{
\item{x}{(vector) a set of values}

\item{p}{(vector) a vector of probability levels.}

\item{na.rm}{(logical) remove NAs.}
}
\value{
A vector of quantiles.
}
\description{
Quantiles by Harrell & Davis 1982 for a vector of probabilities
}
