% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{W}
\alias{W}
\title{Normality index of a sample by the Shapiro test}
\usage{
W(X, index = 1:length(X), ...)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxillary function for bootstrap by 'boot::boot()'
}
