% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibScores.R
\name{zceFun}
\alias{zceFun}
\title{Z-scores based mean absolute calibration error}
\usage{
zceFun(Z, intrv)
}
\arguments{
\item{Z}{(vector) z-scores}

\item{intrv}{(list) intervals generated by `genIntervals`}
}
\value{
The mean abolute calibration error
}
\description{
Z-scores based mean absolute calibration error
}
\examples{
\donttest{
  uE = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E  = rnorm(uE, mean=0, sd=uE)    # Generate errors
  intrv = ErrViewLib::genIntervals(uE, 30)
  zceFun(E/uE, intrv)
}
}
