% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZV.R
\name{ZMSCI}
\alias{ZMSCI}
\title{Estimate statistics of <Z^2>}
\usage{
ZMSCI(
  Z,
  method = c("bootstrap", "stud", "auto"),
  CImethod = c("bca", "perc", "basic"),
  nBoot = 1500,
  level = 0.95,
  parallel = c("no", "multicore")
)
}
\arguments{
\item{Z}{(vector) a data sample}

\item{method}{(string) one of 'auto' (default), 'bootstrap' and 'stud'}

\item{CImethod}{(string) one of 'bca' (default), 'perc' and 'basic'
for the CI estimation algorithm from bootstrap sample}

\item{nBoot}{(integer) number of bootstrap repeats}

\item{level}{(numeric) a probability level for CI (default: 0.95)}

\item{parallel}{(string) one of 'no' (default) and 'multicore'}
}
\value{
A list containing the mean, sd and ci for <Z^2>
}
\description{
Estimate statistics of <Z^2>
}
