% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBS.R
\name{estBS1}
\alias{estBS1}
\title{Title}
\usage{
estBS1(
  error,
  props = c("mue", "mse", "rmsd", "q95", "q95hd", "P1", "W"),
  nboot = 1000,
  do.sip = TRUE,
  eps = 1,
  seed = 123,
  silent = FALSE
)
}
\arguments{
\item{silent}{}
}
\value{

}
\description{
Title
}
