% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genIntervals.R
\name{genIntervals}
\alias{genIntervals}
\title{Generate intervals for local statistics}
\usage{
genIntervals(N, nBin, slide = FALSE)
}
\arguments{
\item{N}{(integer) size of set to partition}

\item{nBin}{(integer) number of contiguous intervals}

\item{slide}{(logical) generate contiguous intervals
(default `slide = FALSE`) or gerenate from a sliding window}
}
\value{
A list with `nbr` the number of intervals, `lwindx` and  `upindx`,
  vectors of indices for the first and last point of each interval.
}
\description{
Generate intervals for local statistics
}
