% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTabStat.R
\name{genTabStat}
\alias{genTabStat}
\title{Generate results table from benchmarking statistics generated
by 'estBS()'}
\usage{
genTabStat(S, comp = TRUE, ref = 0, numDig = 1, units = "a.u.")
}
\arguments{
\item{numDig}{}
}
\value{

}
\description{
Generate results table from benchmarking statistics generated
by 'estBS()'
}
