% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdQuantiles.R
\name{hd}
\alias{hd}
\title{Quantiles estimation by the Harrell & Davis method}
\usage{
hd(x, q = 0.5, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\value{

}
\description{
Quantiles estimation by the Harrell & Davis method
}
