% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAGVZMS.R
\name{plotAGV}
\alias{plotAGV}
\title{Adversarial Group Calibration based on the Scal statistic}
\usage{
plotAGV(
  Z,
  popMin = 50,
  nMC = 100,
  add = FALSE,
  ylim = NULL,
  col = 6,
  control = TRUE,
  colControl = 2,
  dist = c("Normal", "Uniform", "Normp4", "Laplace", "T4"),
  title = "",
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{Z}{(vector) set of z-score values to be tested}

\item{popMin}{(integer) minimal bin count in an interval}

\item{nMC}{(integer) number of random intervals per size and repeats}

\item{add}{(logical) add to previous graph ?}

\item{ylim}{(vector) limits of the y axis}

\item{col}{(integer) color index of main curve}

\item{control}{(logical) estimate AGV for control sample (normal-standard)}

\item{colControl}{(integer) color index of control curve}

\item{dist}{(string) model error distribution to generate the control
values. One of 'Normal' (default), 'Uniform', 'Normp4', 'Laplace' or 'T4'}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}

\item{method}{(string) method used to estimate 95 percent CI on <Z^2>}
}
\value{
Nothing.
}
\description{
Adversarial Group Calibration based on the Scal statistic
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)    # Generate errors
  plotAGV(E/uE)
}
}
