% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalVar.R
\name{plotCalVar}
\alias{plotCalVar}
\title{Plot variance calibration curve (reliability diagram)}
\usage{
plotCalVar(
  X,
  Y,
  logX = FALSE,
  nBin = NULL,
  slide = NULL,
  BSmethod = c("bca", "perc", "basic"),
  nBoot = 0,
  xlim = NULL,
  ylim = NULL,
  title = "",
  label = 0,
  add = FALSE,
  col = 5,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) vector of uncertainties}

\item{Y}{(vector) vector of errors}

\item{logX}{(logical) log-transform X}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{slide}{(logical) use sliding window for subsetting (X,Z)}

\item{BSmethod}{(string) bootstrap variant}

\item{nBoot}{(integer) number of bootstrap replicas}

\item{xlim}{(vector) min and max values of X axis}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{add}{(logical) add to previous graph ?}

\item{col}{(integer) color index of curve to add}

\item{gPars}{(list) graphical parameters}
}
\value{
Used for its plotting side effect.
}
\description{
Plot variance calibration curve (reliability diagram)
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)  # Generate errors
  plotCalVar(uE, E, nBin = 6, nBoot = 500)
}
}
