% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorMat.R
\name{plotCorMat}
\alias{plotCorMat}
\title{Title}
\usage{
plotCorMat(
  X,
  method = "ellipse",
  order = "original",
  main = "",
  label = 0,
  cex.lab = 1,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Title
}
