% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDeltaCDF.R
\name{plotDeltaCDF}
\alias{plotDeltaCDF}
\title{Title}
\usage{
plotDeltaCDF(
  err,
  meth1,
  meth2,
  eps = NULL,
  xmax = NULL,
  xlab = NULL,
  units = "a.u.",
  main = "",
  nboot = 1000,
  label = 0,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Title
}
