% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistHist.R
\name{plotDistHist}
\alias{plotDistHist}
\title{Scatter plot of an error set with marginal histogram}
\usage{
plotDistHist(
  x,
  y,
  uy = NULL,
  nclass = NULL,
  xlab = "x",
  ylab = "y",
  plotGauss = FALSE,
  outLiers = FALSE,
  p = 0.9,
  labels = 1:length(x),
  select = NULL,
  main = NULL,
  plotReg = TRUE,
  plotConf = FALSE,
  plotBA = FALSE,
  plotBAci = FALSE,
  xlim = range(x),
  ylim = range(y),
  scaleLegBA = 0.75,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Scatter plot of an error set with marginal histogram
}
