% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvsPU.R
\name{plotEvsPU}
\alias{plotEvsPU}
\title{Plot error E vs predicted value V or prediction uncertainty uE}
\usage{
plotEvsPU(
  X,
  Y,
  type = c("prop", "horiz"),
  runExt = FALSE,
  runQuant = FALSE,
  runMode = FALSE,
  runMean = FALSE,
  runVar = FALSE,
  cumMAE = FALSE,
  probs = c(0.95),
  xlab = NULL,
  ylab = NULL,
  logX = FALSE,
  logY = FALSE,
  title = NULL,
  myColors = c(5, 4, 3, 1, 2),
  xlim = NULL,
  ylim = NULL,
  scalePoints = 0.5,
  nBin = NULL,
  slide = TRUE,
  label = 0,
  showLegend = TRUE,
  legLoc = "topleft",
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{X}{(vector) prediction uncertainty, uE, or predicted value, V}

\item{Y}{(vector) error or z-score}

\item{type}{(string) type of guide lines ('prop' or 'horiz')}

\item{runExt}{(logical) plot extremal points}

\item{runQuant}{(logical) plot running quantiles (cf. probs)}

\item{runMode}{(logical) plot running mode of distribution}

\item{runMean}{(logical) plot running mean}

\item{cumMAE}{(logical) plot cumulative MAE}

\item{probs}{(vector) probability levels for quantile-based confidence intervals}

\item{xlab}{(string) x axis label}

\item{ylab}{(string) y axis label}

\item{logX}{(logical) log-transform X}

\item{logY}{(logical) log-transform Y after taking absolute value}

\item{title}{(string) a title to display above the plot}

\item{xlim}{(vector) limits of the X axis}

\item{ylim}{(vector) limits of the y axis}

\item{scalePoints}{(numeric) scale factor for points size}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{slide}{(logical) use sliding window for subsetting (X,Y)}

\item{label}{(integer) index of letter for subplot tag}

\item{showLegend}{(logical) display legend}

\item{legLoc}{(string) location of legend (see \link[grDevices]{xy.coord})}

\item{gPars}{(list) graphical parameters}
}
\value{
Plots E vs uE or V
}
\description{
Plot error E vs predicted value V or prediction uncertainty uE
}
\examples{
\donttest{
  uE  = sqrt(rchisq(1000, df = 4))  # Re-scale uncertainty
  E   = rnorm(uE, mean=0, sd=uE)  # Generate errors
  plotEvsPU(uE, E, runQuant = TRUE)
}
}
