% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvsPU.R
\name{plotEvsPU}
\alias{plotEvsPU}
\title{Plot error vs predicted value or prediction uncertainty}
\usage{
plotEvsPU(
  X,
  E,
  type = c("PU", "PV"),
  pExt = FALSE,
  qReg = FALSE,
  xlab = NULL,
  ylab = "Error, E",
  logX = FALSE,
  title = NULL,
  myColors = c(5, 4, 3, 2),
  xlim = NULL,
  ylim = NULL,
  scalePoints = 0.5,
  nBin = NULL,
  slide = TRUE,
  label = 0,
  gPars
)
}
\arguments{
\item{X}{(vector) predicted value or prediction uncertainty}

\item{E}{(vector) error}

\item{type}{(string) type of plot ('PV' or 'PU')}

\item{pExt}{(logical) plot extremal points}

\item{xlab}{(string) x axis label}

\item{ylab}{(string) y axis label}

\item{logX}{(logical) log-transform x}

\item{title}{(string) a title to display above the plot}

\item{xlim}{(vector) limits of the x axis}

\item{ylim}{(vector) limits of the y axis}

\item{scalePoints}{(numeric) scale factor for points size}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{slide}{(logical) use sliding window for subsetting (X,E)}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}

\item{qExt}{(logical) plot quantile regression}
}
\value{
Plots E vs uE.
}
\description{
Plot error vs predicted value or prediction uncertainty
}
