% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLorenz.R
\name{plotLorenz}
\alias{plotLorenz}
\title{Plot of a set of Lorenz curves}
\usage{
plotLorenz(
  X,
  title = "Lorenz curves",
  show.leg = TRUE,
  col.index = 1:ncol(X),
  label = 0,
  leg.lwd = 30,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Plot of a set of Lorenz curves
}
