% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRR.R
\name{plotLRR}
\alias{plotLRR}
\title{Plot local range ratios}
\usage{
plotLRR(
  E,
  U,
  ordX = NULL,
  logX = FALSE,
  prob = c(0.95),
  nBin = NULL,
  plot = TRUE,
  slide = NULL,
  mycols = 1:length(prob),
  xlab = "Calculated value",
  xlim = NULL,
  ylim = NULL,
  title = "",
  legend = "P95",
  legLoc = "bottom",
  legNcol = 3,
  add = FALSE,
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{E}{(vector) Errors}

\item{U}{(matrix) Prediction uncertainties}

\item{ordX}{(vector) set of abscissas to order sample}

\item{logX}{(logical) log-transform abscissas}

\item{prob}{(vector) a set of coverage probabilities for the PUs}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{plot}{(logical) plot the results}

\item{slide}{(logical) use sliding window}

\item{mycols}{(vector) a set of color indexes to gPars colors}

\item{xlab}{(string) abscissa label}

\item{xlim}{(vector) range for abscissa}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{legend}{(string) legend for the dataset}

\item{legLoc}{(string) location of legend (see \link[grDevices]{xy.coord})}

\item{legNcol}{(integer) number of columns for legend}

\item{add}{(logical) add to previous graph ?}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LCP results. Mainly used
  for its plotting side effect.
}
\description{
Plot local range ratios
}
