% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZV.R
\name{plotLZV}
\alias{plotLZV}
\title{Plot local Z-score variance to assess calibration and sharpness}
\usage{
plotLZV(
  X,
  Z,
  logX = FALSE,
  nBin = NULL,
  plot = TRUE,
  slide = NULL,
  nBoot = 1000,
  xlab = "Calculated value",
  xlim = NULL,
  ylim = NULL,
  title = "",
  legLoc = "bottom",
  legNcol = 3,
  label = 0,
  gPars = NULL
)
}
\arguments{
\item{nBin}{(integer) number of intervals for local coverage stats}

\item{plot}{(logical) plot the results}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}

\item{Data}{(data.frame) dataframe with predictor(s) and reference}

\item{uP}{(vector) a set of prediction uncertainties}

\item{legloc}{(string) location of the legend (default: "bottom")}
}
\value{
Invisibly returns a list of LZV results. Mainly used
  for its plotting side effect.
}
\description{
Plot local Z-score variance to assess calibration and sharpness
}
