% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParallel.R
\name{plotParallel}
\alias{plotParallel}
\title{Parallel plot of datasets with outliers detection features}
\usage{
plotParallel(
  X,
  maxPoints = nrow(X),
  labels = NULL,
  lab.thresh = 0,
  colors = NULL,
  rescale = TRUE,
  scramble = FALSE,
  outliers = FALSE,
  outLabCex = 1,
  outMethCex = 1,
  xlim = c(1, ncol(X)),
  ylim = NULL,
  units = "a.u.",
  ylab = "Errors",
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Parallel plot of datasets with outliers detection features
}
