% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcoverage.R
\name{plotPcoverage}
\alias{plotPcoverage}
\title{Plot local coverage probabilities to assess calibration and sharpness}
\usage{
plotPcoverage(
  Data,
  corTrend = FALSE,
  fo = NA,
  ordX = NULL,
  logX = FALSE,
  CImeth = c("eq", "pred", "dist"),
  prob = c(0.5, 0.75, 0.95),
  dist = c("norm", "t"),
  shape = 2,
  valid = c("kfold", "loo"),
  nFold = 10,
  nRepeat = 10,
  nBin = NULL,
  binomCI = c("wilson", "wilsoncc", "clopper-pearson", "agresti-coull", "jeffreys"),
  plot = TRUE,
  slide = NULL,
  mycols = 1:length(prob),
  xlab = "Calculated value",
  xlim = NULL,
  ylim = c(0, 1),
  title = "",
  legLoc = "bottom",
  legNcol = 3,
  label = 0,
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{Data}{(data.frame) dataframe with predictor(s) and reference}

\item{corTrend}{(logical) flag to correct trend}

\item{fo}{(formula) trend correction formula}

\item{ordX}{(vector) set of abscissas to order sample}

\item{logX}{(logical) log-transform abscissas}

\item{CImeth}{(string) method to estimate CI limits}

\item{prob}{(vector) a set of coverage probabilities to test}

\item{dist}{(string) a distribution}

\item{shape}{(numeric) shape parameter (> 0, maybe non-integer).}

\item{valid}{(string) a cross-validation method: "kfold" (default) or "loo"}

\item{nFold}{(integer) number of folds (default: 10)}

\item{nRepeat}{(integer) number of repeats for k-fold cross-validation}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{binomCI}{(string) name of method to estimate Binomial Proportion CI}

\item{plot}{(logical) plot the results}

\item{slide}{(logical) use sliding window}

\item{mycols}{(vector) a set of color indexes to gPars colors}

\item{xlab}{(string) abscissa label}

\item{xlim}{(vector) range for abscissa}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{legLoc}{(string) location of legend (see \link[grDevices]{xy.coord})}

\item{legNcol}{(integer) number of columns for legend}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Invisibly returns a list of LCP results. Mainly used
  for its plotting side effect.
}
\description{
Plot local coverage probabilities to assess calibration and sharpness
}
