% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcoverage.R
\name{plotPcoverage}
\alias{plotPcoverage}
\title{Plot local coverage probabilities to assess calibration and sharpness}
\usage{
plotPcoverage(
  R,
  C,
  uC = NULL,
  corTrend = FALSE,
  degree = 1,
  CImeth = c("eq", "pred", "dist"),
  prob = c(0.5, 0.75, 0.95),
  dist = c("norm", "t"),
  shape = 2,
  mycols = 1:length(prob),
  valid = c("kfold", "loo"),
  nFold = 10,
  nRepeat = 10,
  nBin = 10,
  xlab = "Calculated value",
  ylim = NA,
  title = "",
  legloc = "bottom",
  label = 0,
  gPars
)
}
\arguments{
\item{R}{(vector) a reference dataset}

\item{C}{(vector) a set of predicted values}

\item{uC}{(vector) a set of prediction uncertainties}

\item{corTrend}{(logical) flag to correct trend}

\item{degree}{(integer) polynomial degree of trend}

\item{CImeth}{(string) method to estimate CI limits}

\item{prob}{(vector) a set of coverage probabilities to test}

\item{dist}{(string) a distribution}

\item{shape}{(numeric) shape parameter (> 0, maybe non-integer).}

\item{mycols}{(vector) a set of color indexes to gPars colors}

\item{valid}{(string) a cross-validation method: "kfold" (default) or "loo"}

\item{nFold}{(integer) number of folds (default: 10)}

\item{nRepeat}{(integer) number of repeats for k-fold cross-validation}

\item{nBin}{(integer) number of intervals for local coverage stats}

\item{ylim}{(vector) limits of the y axis}

\item{title}{(string) a title to display above the plot}

\item{legloc}{(string) location of the legend (default: "bottom")}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
Nothing. It is used for its side effect (plot).
}
\description{
Plot local coverage probabilities to assess calibration and sharpness
}
