% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPIT.R
\name{plotPIT}
\alias{plotPIT}
\title{Plot PIT histograms}
\usage{
plotPIT(
  Z,
  dist = c("norm", "t"),
  shape = 2,
  col = 5,
  breaks = 10,
  title = "",
  label = 0,
  gPars
)
}
\arguments{
\item{Z}{(vector) a set of z-zscores}

\item{dist}{(string) a distribution}

\item{shape}{(numeric) shape parameter (> 0, maybe non-integer).}

\item{col}{(integer) a color number for the histigram}

\item{breaks}{(integer) number of breaks in the histogram}

\item{title}{(string) a title for the plot}

\item{label}{(integer) index of letter for subplot tag}

\item{gPars}{(list) graphical parameters}
}
\value{
a plot
}
\description{
Plot PIT histograms
}
