% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPpnorm.R
\name{plotPpnorm}
\alias{plotPpnorm}
\title{plotPpnorm}
\usage{
plotPpnorm(X, title = "", plotCI = FALSE, score = TRUE, gPars)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Deviation of ECDF from normality
}
