% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUncEcdf.R
\name{plotUncEcdf}
\alias{plotUncEcdf}
\title{Title}
\usage{
plotUncEcdf(
  X,
  xlab = NULL,
  xmax = NULL,
  title = "",
  show.leg = TRUE,
  show.MAE = FALSE,
  show.Q95 = TRUE,
  Q.algo = "HD",
  col.index = 1:ncol(X),
  weights = NULL,
  units = "a.u.",
  label = 0,
  leg.lwd = 30,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Title
}
