% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZscoreQqnorm.R
\name{plotZscoreQqnorm}
\alias{plotZscoreQqnorm}
\title{plotZscoreQqnorm}
\usage{
plotZscoreQqnorm(
  R,
  sig,
  lim = NULL,
  plotCI = FALSE,
  plorReg = FALSE,
  markOutliers = FALSE,
  title = "",
  gPars
)
}
\arguments{
\item{markOutliers}{}
}
\value{

}
\description{
plotZscoreQqnorm
}
