% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZscoreQqnorm.R
\name{plotZscoreQqnorm}
\alias{plotZscoreQqnorm}
\title{plotZscoreQqnorm}
\usage{
plotZscoreQqnorm(
  R,
  sig,
  lim = NULL,
  plotCI = FALSE,
  plotReg = FALSE,
  markOutliers = FALSE,
  labels = 1:length(R),
  title = "",
  gPars = ErrViewLib::setgPars()
)
}
\arguments{
\item{R}{-}

\item{sig}{-}

\item{lim}{-}

\item{plotCI}{-}

\item{plotReg}{-}

\item{markOutliers}{-}

\item{labels}{-}

\item{title}{-}

\item{gPars}{-}
}
\value{

}
\description{
plotZscoreQqnorm
}
