% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcoverage.R
\name{predQ}
\alias{predQ}
\title{Auxillary function for plotPcoverage}
\usage{
predQ(
  cL,
  rL,
  cT,
  rT,
  corTrend = FALSE,
  fo = NA,
  prob = c(0.5, 0.75, 0.95),
  CImeth = c("eq", "pred", "dist"),
  dist = c("norm", "t"),
  shape = 2
)
}
\arguments{
\item{CImeth}{}
}
\value{

}
\description{
Auxillary function for plotPcoverage
}
