% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalCurve.R
\name{predQ}
\alias{predQ}
\title{Auxillary function for plotCalCurve
Predicted quantiles of errors}
\usage{
predQ(
  Learn,
  Test,
  corTrend = FALSE,
  fo = NA,
  prob = seq(0.005, 0.995, by = 0.005),
  CImeth = c("eq", "pred", "dist"),
  dist = c("norm", "t"),
  shape = 2
)
}
\arguments{
\item{corTrend}{-}

\item{fo}{-}

\item{prob}{-}

\item{CImeth}{-}

\item{cLearn}{-}

\item{rLearn}{-}

\item{cTest}{-}

\item{rTest}{-}
}
\value{
a list
}
\description{
Auxillary function for plotCalCurve
Predicted quantiles of errors
}
