% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcoverage.R
\name{predQ}
\alias{predQ}
\title{Auxillary function for plotPcoverage}
\usage{
predQ(
  cLearn,
  rLearn,
  cTest,
  rTest,
  ucLearn = NULL,
  ucTest = NULL,
  corTrend = FALSE,
  fo = NA,
  prob = c(0.5, 0.75, 0.95),
  qreg = FALSE
)
}
\arguments{
\item{qreg}{}
}
\value{

}
\description{
Auxillary function for plotPcoverage
}
