% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGpars.R
\name{setgPars}
\alias{setgPars}
\title{Set default graphical parameters}
\usage{
setgPars(type = c("shiny", "plot"), gPars = list())
}
\arguments{
\item{type}{(string) plot type. Default: 'shiny'}

\item{gPars}{(list) list of graphical parameters to supersede defaults}
}
\value{

}
\description{
Set default graphical parameters
}
