% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLZV.R
\name{varZCI}
\alias{varZCI}
\title{Estimate statistics of the variance of a sample}
\usage{
varZCI(
  Z,
  method = c("bootstrap", "cho", "chisq", "auto"),
  CImethod = c("bca", "perc", "basic", "stud"),
  nBoot = 5000,
  level = 0.95,
  parallel = FALSE,
  cl = NULL
)
}
\arguments{
\item{Z}{(vector) a data sample}

\item{method}{(string) one of 'auto' (default), 'bootstrap', 'cho' and
'chisq' for the estimation of the statistics}

\item{CImethod}{(string) one of 'bca' (default), 'perc' and 'basic'
for the CI estimation algorithm from bootstrap sample}

\item{nBoot}{(integer) number of bootstrap repeats}

\item{level}{(numeric) a probability level for CI (default: 0.95)}

\item{parallel}{(logical) use parallelized bootstrap (default: `FALSE`)}

\item{cl}{(object) a cluster object for parallelized bootstrap (default: `NULL`)}
}
\value{
A list containing the mean, sd and ci for the variance
    of the Z sample
}
\description{
Estimate statistics of the variance of a sample
}
