% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Zanardi}
\alias{Zanardi}
\title{Zanardi index (Clementi2019)}
\usage{
Zanardi(X, indx = 1:length(X), ...)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Auxiliary function for bootstrap by 'boot::boot()'
}
